Migrate UCC-variant tests
