+++ test/algorithms/ground_state_solvers/test_advanced_ucc_variants.py
@@ -1,0 +1,141 @@
+# This code is part of Qiskit.
+#
+# (C) Copyright IBM 2019, 2021.
+#
+# This code is licensed under the Apache License, Version 2.0. You may
+# obtain a copy of this license in the LICENSE.txt file in the root directory
+# of this source tree or at http://www.apache.org/licenses/LICENSE-2.0.
+#
+# Any modifications or derivative works of this code must retain this
+# copyright notice, and modified files need to carry a notice indicating
+# that they have been altered from the originals.
+
+""" Test of UCCSD and HartreeFock extensions """
+
+import unittest
+
+from test import QiskitNatureTestCase
+
+from qiskit import BasicAer
+from qiskit.utils import QuantumInstance
+from qiskit.algorithms import VQE
+from qiskit.algorithms.optimizers import SLSQP
+from qiskit.test import slow_test
+
+from qiskit_nature.algorithms.ground_state_solvers import GroundStateEigensolver
+from qiskit_nature.circuit.library import HartreeFock
+from qiskit_nature.circuit.library.ansatzes import SUCCD, PUCCD
+from qiskit_nature.drivers import PySCFDriver
+from qiskit_nature.mappers.second_quantization import ParityMapper
+from qiskit_nature.operators.second_quantization.qubit_converter import QubitConverter
+from qiskit_nature.problems.second_quantization.molecular import MolecularProblem
+from qiskit_nature.transformers import FreezeCoreTransformer
+
+# pylint: disable=invalid-name
+
+
+class TestUCCSDHartreeFock(QiskitNatureTestCase):
+    """Test for these extensions."""
+
+    def setUp(self):
+        super().setUp()
+        self.driver = PySCFDriver(atom='H 0 0 0.735; H 0 0 0', basis='631g')
+
+        self.qubit_converter = QubitConverter(ParityMapper(), two_qubit_reduction=True)
+
+        self.molecular_problem = MolecularProblem(self.driver, [FreezeCoreTransformer()])
+
+        self.num_spin_orbitals = 8
+        self.num_particles = (1, 1)
+
+        self.reference_energy_pUCCD = -1.1434447924298028
+        self.reference_energy_UCCD0 = -1.1476045878481704
+        self.reference_energy_UCCD0full = -1.1515491334334347
+        # reference energy of UCCSD/VQE with tapering everywhere
+        self.reference_energy_UCCSD = -1.1516142309717594
+        # reference energy of UCCSD/VQE when no tapering on excitations is used
+        self.reference_energy_UCCSD_no_tap_exc = -1.1516142309717594
+        # excitations for succ
+        self.reference_singlet_double_excitations = [[0, 1, 4, 5], [0, 1, 4, 6], [0, 1, 4, 7],
+                                                     [0, 2, 4, 6], [0, 2, 4, 7], [0, 3, 4, 7]]
+        # groups for succ_full
+        self.reference_singlet_groups = [[[0, 1, 4, 5]], [[0, 1, 4, 6], [0, 2, 4, 5]],
+                                         [[0, 1, 4, 7], [0, 3, 4, 5]], [[0, 2, 4, 6]],
+                                         [[0, 2, 4, 7], [0, 3, 4, 6]], [[0, 3, 4, 7]]]
+
+    @slow_test
+    def test_uccsd_hf_qpUCCD(self):
+        """ paired uccd test """
+        optimizer = SLSQP(maxiter=100)
+
+        initial_state = HartreeFock(self.num_spin_orbitals,
+                                    self.num_particles,
+                                    self.qubit_converter)
+
+        var_form = PUCCD(self.qubit_converter,
+                         self.num_particles,
+                         self.num_spin_orbitals,
+                         initial_state=initial_state)
+
+        solver = VQE(var_form=var_form, optimizer=optimizer,
+                     quantum_instance=QuantumInstance(
+                         backend=BasicAer.get_backend('statevector_simulator')))
+
+        gsc = GroundStateEigensolver(self.qubit_converter, solver)
+
+        result = gsc.solve(self.molecular_problem)
+
+        self.assertAlmostEqual(result.total_energies[0], self.reference_energy_pUCCD, places=6)
+
+    @slow_test
+    def test_uccsd_hf_qUCCD0(self):
+        """ singlet uccd test """
+        optimizer = SLSQP(maxiter=100)
+
+        initial_state = HartreeFock(self.num_spin_orbitals,
+                                    self.num_particles,
+                                    self.qubit_converter)
+
+        var_form = SUCCD(self.qubit_converter,
+                         self.num_particles,
+                         self.num_spin_orbitals,
+                         initial_state=initial_state)
+
+        solver = VQE(var_form=var_form, optimizer=optimizer,
+                     quantum_instance=QuantumInstance(
+                         backend=BasicAer.get_backend('statevector_simulator')))
+
+        gsc = GroundStateEigensolver(self.qubit_converter, solver)
+
+        result = gsc.solve(self.molecular_problem)
+
+        self.assertAlmostEqual(result.total_energies[0], self.reference_energy_UCCD0, places=6)
+
+    @unittest.skip("Skip until https://github.com/Qiskit/qiskit-nature/issues/91 is closed.")
+    def test_uccsd_hf_qUCCD0full(self):
+        """ singlet full uccd test """
+        optimizer = SLSQP(maxiter=100)
+
+        initial_state = HartreeFock(self.num_spin_orbitals,
+                                    self.num_particles,
+                                    self.qubit_converter)
+
+        # TODO: add `full` option
+        var_form = SUCCD(self.qubit_converter,
+                         self.num_particles,
+                         self.num_spin_orbitals,
+                         initial_state=initial_state)
+
+        solver = VQE(var_form=var_form, optimizer=optimizer,
+                     quantum_instance=QuantumInstance(
+                         backend=BasicAer.get_backend('statevector_simulator')))
+
+        gsc = GroundStateEigensolver(self.qubit_converter, solver)
+
+        result = gsc.solve(self.molecular_problem)
+
+        self.assertAlmostEqual(result.total_energies[0], self.reference_energy_UCCD0full, places=6)
+
+
+if __name__ == '__main__':
+    unittest.main()
